refactor: move the context to a separate file
