+++ src/shared/lib/theme/context.ts
@@ -1,0 +1,7 @@
+import { createContext } from "react";
+import type { ThemeContextType } from "./model/types";
+
+export const ThemeContext = createContext<ThemeContextType>({
+  theme: "light",
+  toggleTheme: () => console.warn("Theme provider not initialized!"),
+});
