+++ src/shared/lib/theme/model/hooks/useTheme.ts
@@ -1,5 +1,5 @@
 import { useContext } from "react";
-import { ThemeContext } from "./ThemeProvider";
+import { ThemeContext } from "../../context";
 
 export function useTheme() {
   return useContext(ThemeContext);
