+++ src/shared/lib/theme/model/types.ts
